Add role assignment service checks
Also generalise util functions (non-user specific)
