+++ api/user/utils.ts
@@ -1,4 +1,5 @@
 import { UserInfo } from '../auth/interfaces/UserInfo';
+import { allDoNotContainDangerousCharacters, urlHasNoDangerousCharacters } from '../utils';
 import { CASE_ALLOCATOR_ROLE, ORGANISATION_ROLE_TYPE } from './constants';
 import { RoleAssignment } from './interfaces/roleAssignment';
 
@@ -99,28 +100,6 @@ export function userDetailsValid(userInfo: UserInfo): boolean {
     return false;
   }
   return true;
-}
-
-export function allDoNotContainDangerousCharacters(values: string[]) {
-  for (const value of values) {
-    if (!hasNoDangerousCharacters(value)) {
-      // if one value contains dangerous characters, return false
-      return false;
-    }
-  }
-  return true;
-}
-
-export function hasNoDangerousCharacters(value: string): boolean {
-  // the below only checks for the characters in the string
-  // return /^[^%<>^$//]+$/.test(value);
-  // substring approach below preferred
-  return /^(?!.*\/\*|.*\/\/|.*;|.*&|.*\?|.*<|.*\^|.*>).+$/.test(value);
-}
-
-// url may have special characters but should guard against other dangerous characters
-export function urlHasNoDangerousCharacters(value: string): boolean {
-  return /^(?!.*\/\*|.*;|.*<|.*\^|.*>).+$/.test(value);
 }
 
 export function hasRoleCategory(roles: string[], roleName: string): boolean {
